Fix VC++ compilation warnings and errors

+++ src/strsmush.h
@@ -7,8 +7,8 @@
 
 class StringSmusher {
  public:
-    static int amount(std::wstring const&, std::wstring const&, wchar_t, uint32_t);
-    static std::wstring smush(std::wstring const&, std::wstring const&, int, wchar_t, uint32_t);
+    static size_t amount(std::wstring const&, std::wstring const&, wchar_t, uint32_t);
+    static std::wstring smush(std::wstring const&, std::wstring const&, size_t, wchar_t, uint32_t);
 };
 
 #endif  // SRC_STRSMUSH_H_
